Make Block#add(...) return Values instead of AddExpressions
Originally, AddExpression was needed to make sure unused values are not
emitted in bytecode. This has been since replaced with proper value usage
tracking, which does not require the developer to do anything.

Since most uses of AddExpression were just calls to value() without
arguments, I decided to remove it and add a couple of helpers for (named)
variables.