Add a parser generator (WIP)
When all you have is a code generation library, all problems start to look like
compilers. In this case, I built a parser generator that uses code4jvm as backend.
It is still totally untested and probably has horrible bugs.

The generator creates LL parsers, but LL(1) is not... strictly required. If
a predict set has multiple possibilities, the parser will just try them all.
Likewise, if parsing fails, backtracking is supported. I hope this will be
useful when thinking about error recovery.

Incremental parsing is a can of worms that I haven't touched yet. In theory,
all the building blocks for it are in place... in practice, it is not so easy.

Other changes:
* Tokens are no longer extended; instead, user code should define token types
* Code4jvm support for null constants