Add support for virtual nodes
Previous efforts to generate smaller code were not enough. I had to finally
bite the bullet and add support for separate virtual nodes - that is,
nodes that would exist in parse tree if we had one but should not exist in the
AST parserkit produces.

Hardest part of this was probably to keep left recursion elimination working.