lua4jvm: Function declarations, calls and specialization
Lua functions are compiled when they are called, not when they are declared.
This way, the compiler can consider types of each call site to generate
specializations. This allows e.g. using primitive types in generated bytecode
to improve performance, as well as opens doors to other optimizations in future.

This took quite a while to get right. Compared to earlier attempts, this:
* Uses 'advanced' MethodHandle features to keep call sites as simple
* Uses classData (relatively recent hidden class feature) and dynamic constants
  instead of generating unnecessary fields in function specializations
* Fully separates types from values of upvalues