lua4jvm: Change parser to Antlr 4
My own LL parser (parserkit) is slow, buggy, and lacks important features such
as error recovery. It was certainly a learning experience, but not a good
starting point for a reliable language implementation. I'd much prefer working
on the missing Lua features and further runtime optimizations.

The Lua grammar is based on Lua 5.4 spec and inspired by official Lua
grammar. I chose not to use the official grammar, because supporting other
Lua versions or forks like Luau is out-of-scope at the moment.

The grammar still needs further testing, but is at least not worse than what
I had before!