lua4jvm: Add support for varargs (multi-values)
Finally! This is essentially the second attempt for supporting varargs.
The first one stalled due to my poor understanding of the topic and the
limitations of the function linker at that time.

For former, I learned many *fascinating* details from this great blog post:
https://benaiah.me/posts/everything-you-didnt-want-to-know-about-lua-multivals

For latter, well, the linker was heavily refactored when optimized table
access was added. It no longer has weird special cases around usage of runtime
types or such things. Still, adding varargs (=multivals in lua4jvm) was not
exactly simple:

* Multivals are stored as Object[] AND any call target might return one
  * ... however, not all call sites will accept more than the first value
  * => specialize call targets to avoid unnecessary array creation
* Handling of multivals is not free
  * => don't do it if call target has been specialized not to return one
  * Several fun special cases Java-Lua FFI
* Multivals can be used as function arguments (spreading them)!
* Multivals can be returned (spreading them)!

Yeah, fun.