lua4jvm: Add a coroutine implementation
Implementing Lua's coroutines on top of threads is actually not THAT difficult.
With virtual threads, even the performance characteristics of it should be
similar enough to PUC Lua. And since JVM does support multithreading, we don't
actually need to stop Java code from running in threads that yield!

Lua API pending.